refactor fetch and sleeper into single package
